fix: Address code duplication
Remove most of the code in addEvent and editEvent.
They know both call a composable BaseEventScreen which
basically call everything they plus has a callback for
additional content the other one doesn't have.
