+++ app/src/main/java/com/swent/mapin/ui/event/AddEventScreen.kt
@@ -1,25 +1,16 @@
 package com.swent.mapin.ui.event
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.ime
-import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
@@ -129,82 +120,30 @@ fun AddEventScreen(
     onDone: () -> Unit = {},
 ) {
   val formState = rememberEventFormState()
-  val locations by locationViewModel.locations.collectAsState()
-  val scrollState = rememberScrollState()
 
-  val errorFields =
-      formState.getErrorFields(
-          titleFieldName = stringResource(R.string.title_field),
-          dateFieldName = stringResource(R.string.date_field),
-          timeFieldName = stringResource(R.string.time),
-          locationFieldName = stringResource(R.string.location_field),
-          descriptionFieldName = stringResource(R.string.description_field),
-          tagFieldName = stringResource(R.string.tag_field),
-          priceFieldName = stringResource(R.string.price_field))
-
-  Scaffold(contentWindowInsets = WindowInsets.ime) { padding ->
-    Column(modifier = modifier.padding(padding).fillMaxWidth().navigationBarsPadding()) {
-      // TopBar
-      EventTopBar(
-          title = "New Event",
-          testTags = AddEventScreenTestTags,
-          isEventValid = formState.isValid(),
-          onCancel = onCancel,
-          onSave = {
-            formState.validateAllFields()
-
-            if (!formState.isValid()) return@EventTopBar
-
-            val timestamps = formState.parseTimestamps() ?: return@EventTopBar
-
-            val (startTs, endTs) = timestamps
-
-            saveEvent(
-                eventViewModel,
-                formState.title.value,
-                formState.description.value,
-                formState.gotLocation.value,
-                startTs,
-                endTs,
-                Firebase.auth.currentUser?.uid,
-                extractTags(formState.tag.value),
-                formState.isPublic.value,
-                onDone,
-                formState.price.value.toDoubleOrNull() ?: 0.0)
-          })
-
-      // Prominent validation banner shown right after the top bar when user attempted to save
-      if (formState.showValidation.value && !formState.isValid()) {
-        ValidationBanner(errorFields, AddEventScreenTestTags)
-      }
-
-      Column(modifier = Modifier.fillMaxWidth().verticalScroll(scrollState)) {
-        Spacer(modifier = Modifier.padding(5.dp))
-
-        EventFormBody(
-            title = formState.title,
-            titleError = formState.titleError,
-            date = formState.date,
-            dateError = formState.dateError,
-            time = formState.time,
-            timeError = formState.timeError,
-            endDate = formState.endDate,
-            endDateError = formState.endDateError,
-            endTime = formState.endTime,
-            endTimeError = formState.endTimeError,
-            location = formState.location,
-            locationError = formState.locationError,
-            locations = locations,
-            gotLocation = formState.gotLocation,
-            locationExpanded = formState.locationExpanded,
-            locationViewModel = locationViewModel,
-            description = formState.description,
-            descriptionError = formState.descriptionError,
-            tag = formState.tag,
-            tagError = formState.tagError,
-            testTags = AddEventScreenTestTags)
-
-        Spacer(modifier = Modifier.padding(bottom = 10.dp))
+  BaseEventScreen(
+      modifier = modifier,
+      formState = formState,
+      locationViewModel = locationViewModel,
+      title = "New Event",
+      testTags = AddEventScreenTestTags,
+      onCancel = onCancel,
+      onCommit = { startTs, endTs ->
+        saveEvent(
+            eventViewModel,
+            formState.title.value,
+            formState.description.value,
+            formState.gotLocation.value,
+            startTs,
+            endTs,
+            Firebase.auth.currentUser?.uid,
+            extractTags(formState.tag.value),
+            formState.isPublic.value,
+            onDone,
+            formState.price.value.toDoubleOrNull() ?: 0.0)
+      },
+      extraContent = {
+        Spacer(modifier = Modifier.padding(bottom = 5.dp))
 
         // Price field
         Text(
@@ -231,24 +170,19 @@ fun AddEventScreen(
         Spacer(modifier = Modifier.padding(bottom = 5.dp))
 
         // Public/Private switch
-        if (formState.isPublic.value) {
-          PublicSwitch(
-              isPublic = formState.isPublic.value,
-              onPublicChange = { formState.isPublic.value = it },
-              "This event will be public",
-              "Others can see this event on the map",
-              Modifier.testTag(AddEventScreenTestTags.PUBLIC_SWITCH),
-              Modifier.testTag(AddEventScreenTestTags.PUBLIC_TEXT))
-        } else {
-          PublicSwitch(
-              isPublic = formState.isPublic.value,
-              onPublicChange = { formState.isPublic.value = it },
-              "This event will be private",
-              "Others will not see this event on the map",
-              Modifier.testTag(AddEventScreenTestTags.PUBLIC_SWITCH),
-              Modifier.testTag(AddEventScreenTestTags.PUBLIC_TEXT))
-        }
-      }
-    }
-  }
+        val publicTitle =
+            if (formState.isPublic.value) "This event will be public"
+            else "This event will be private"
+        val publicSubtitle =
+            if (formState.isPublic.value) "Others can see this event on the map"
+            else "Others will not see this event on the map"
+
+        PublicSwitch(
+            isPublic = formState.isPublic.value,
+            onPublicChange = { formState.isPublic.value = it },
+            title = publicTitle,
+            subTitle = publicSubtitle,
+            modifier = Modifier.testTag(AddEventScreenTestTags.PUBLIC_SWITCH),
+            textTestTag = Modifier.testTag(AddEventScreenTestTags.PUBLIC_TEXT))
+      })
 }
